Stop the config watcher on shutdown and remove dead code
- Register an OnShutdown hook that stops the Consul watch plan, so a
  CoreDNS reload no longer leaks a goroutine that keeps polling Consul.
- Remove the unused records.HandleRecord stub and rename its file to
  record.go, leaving only the Record type and GetRecordTTL.
- Drop the duplicate GetDefaultTTL in favour of records.GetRecordTTL.
- Skip the Consul integration test when the server is unreachable and
  remove the no-op logger restore, so the suite passes without Consul.
